Refactor: Move styles to components and pages folders
